REG-42 Reinstated repository-hbase module due to dependency from Play app on the loader package.  Kept separate hbase-loader module for fat jar assembly which now just has the BuildLoaderApp main class
