+++ build.sbt
@@ -6,21 +6,18 @@ lazy val `sbr-admin-data` = (project in file("."))
   .enablePlugins(BuildInfoPlugin, PlayScala)
   .disablePlugins(AssemblyPlugin)
   .dependsOn(model)
-  .dependsOn(`hbase-connector`)
-  .aggregate(model, `hbase-connector`, `hbase-loader`)
+  .dependsOn(`repository-hbase`)
+  .aggregate(model, `repository-hbase`, `hbase-loader`)
 
 lazy val model = project
   .disablePlugins(AssemblyPlugin)
 
-//lazy val `repository-hbase` = project
-//  .dependsOn(model)
-
-lazy val `hbase-connector` = project
+lazy val `repository-hbase` = project
   .disablePlugins(AssemblyPlugin)
   .dependsOn(model)
 
 lazy val `hbase-loader` = project
-  .dependsOn(`hbase-connector`)
+  .dependsOn(`repository-hbase`)
 
 /**
 * SETTINGS AND CONFIGURATION
