--- hbase-connector/src/test/resources/application.conf
@@ -1,35 +1,0 @@
-admin-data-repository {
-  executor = "thread-pool-executor"
-  type = PinnedDispatcher
-}
-
-hbase {
-  authentication.username = "username"
-  authentication.username = ${?HBASE_AUTHENTICATION_USERNAME}
-
-  authentication.password = "password"
-  authentication.password = ${?HBASE_AUTHENTICATIONM_PASSWORD}
-
-  rest.endpoint = "http://localhost:8080"
-  rest.endpoint = ${?HBASE_REST_ENDPOINT}
-
-  column.family = "d"
-  column.family = ${?HBASE_REST_COLUMN_FAMILY}
-
-  namespace = ""
-  namespace = ${?HBASE_REST_NAMESPACE}
-
-  table.name = "data"
-  table.name = ${?HBASE_REST_TABLE_NAME}
-
-//  namespace = "sbr_local_db"
-//  namespace = ${?HBASE_REST_NAMESPACE}
-//
-//  table.name = "enterprise"
-//  table.name = ${?HBASE_REST_TABLE_NAME}
-}
-
-ws {
-  request.timeout = 15000
-  request.timeout = ${?SBR_EXTERNAL_REQUEST_TIMEOUT}
-}
