--- hbase-loader/src/main/java/hbase/load/CSVDataKVMapper.java
@@ -1,163 +1,0 @@
-package hbase.load;
-
-import au.com.bytecode.opencsv.CSVParser;
-import hbase.model.AdminData;
-import hbase.util.RowKeyUtils;
-import org.apache.hadoop.hbase.KeyValue;
-import org.apache.hadoop.hbase.client.Put;
-import org.apache.hadoop.hbase.io.ImmutableBytesWritable;
-import org.apache.hadoop.io.LongWritable;
-import org.apache.hadoop.io.Text;
-import org.apache.hadoop.mapreduce.Mapper;
-import org.joda.time.format.DateTimeFormat;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import org.joda.time.YearMonth;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static hbase.load.BulkLoader.REFERENCE_PERIOD;
-import static org.apache.hadoop.hbase.util.Bytes.toBytes;
-
-/**
- * hbase org.apache.hadoop.hbase.mapreduce.ImportTsv -Dimporttsv.separator=, -Dimporttsv.mapper.class=my.Mapper
- * <p>
- * hbase jar /Users/harrih/Git/sbr-hbase-connector/target/sbr-hbase-connector-1.0-SNAPSHOT-distribution.jar
- */
-public class CSVDataKVMapper extends
-        Mapper<LongWritable, Text, ImmutableBytesWritable, Put> {
-
-    private static final String COLUMN_HEADINGS = "csv.column.headings";
-    static final String ROWKEY_POSITION = "csv.id.position";
-    public static final String HEADER_STRING = "csv.header.string";
-    private static final Logger LOG = LoggerFactory.getLogger(CSVDataKVMapper.class.getName());
-    private static final byte[] LAST_UPDATED_BY_COLUMN = toBytes("updatedBy");
-    private static final byte[] LAST_UPDATED_BY_VALUE = toBytes("Data Load");
-    private static final String COMMA = ",";
-
-    private CSVParser csvParser;
-    private YearMonth referencePeriod;
-    private byte[][] columnNames;
-    private final byte[] columnFamily = toBytes("d");
-    private int rowKeyFieldPosition;
-
-    private String getHeaderString() {
-        return System.getProperty(HEADER_STRING, "");
-    }
-
-    private int getRowKeyFieldPosition() {
-        return Integer.valueOf(System.getProperty(ROWKEY_POSITION, "1"));
-    }
-
-    private boolean useCsvHeaderAsColumnNames() {
-        return System.getProperty(COLUMN_HEADINGS, "").isEmpty();
-    }
-
-    private byte[][] getColumnNames() {
-        String[] headings = System.getProperty(COLUMN_HEADINGS).split(COMMA);
-        byte[][] columnNames = new byte[headings.length][];
-        int i = 0;
-        for (String heading : headings) {
-            columnNames[i++] = toBytes(heading);
-        }
-        return columnNames;
-    }
-
-    @Override
-    protected void setup(Context context) throws IOException, InterruptedException {
-        rowKeyFieldPosition = getRowKeyFieldPosition();
-        LOG.debug("Id field is a position {} in CSV file", rowKeyFieldPosition);
-        if (!useCsvHeaderAsColumnNames()) {
-            LOG.debug("Using supplied column headers for file not those in the csv file");
-            columnNames = getColumnNames();
-        }
-        String periodStr = System.getProperty(REFERENCE_PERIOD);
-        try {
-            referencePeriod = YearMonth.parse(periodStr, DateTimeFormat.forPattern(AdminData.REFERENCE_PERIOD_FORMAT()));
-        } catch (Exception e) {
-            LOG.error("Cannot parse '{}' system property with value '{}'. Format should be '{}'", REFERENCE_PERIOD, periodStr, AdminData.REFERENCE_PERIOD_FORMAT());
-            throw e;
-        }
-        if (getHeaderString().isEmpty() && useCsvHeaderAsColumnNames()){
-            LOG.error("If no header row identifying string is specified then column heading must be supplied");
-            throw new IllegalArgumentException("System property not set " + HEADER_STRING);
-        }
-        csvParser = new CSVParser();
-    }
-
-    @Override
-    protected void map(LongWritable key, Text value, Context context) throws InterruptedException, IOException {
-        // Skip blank rows
-        if (value == null || value.getLength() == 0) return;
-
-        // Skip header
-        if (isHeaderRow(value)) return;
-
-        String[] fields = parseLine(value, context);
-        if (fields == null) return;
-
-        // Key: e.g. "201706~07382019"
-        String rowKeyStr = RowKeyUtils.createRowKey(referencePeriod, fields[rowKeyFieldPosition]);
-        writeRow(context, rowKeyStr, fields);
-    }
-
-    private boolean isHeaderRow(Text value) throws IOException {
-        String headerString = getHeaderString();
-        if (!headerString.isEmpty()) {
-            if (value.find(headerString) > -1) {
-                if (useCsvHeaderAsColumnNames()) {
-                    LOG.debug("Found csv header row: {}", value.toString());
-                    LOG.debug("Using header row as table column names");
-                    try {
-                        String[] columnNameStrs = csvParser.parseLine(value.toString().trim());
-                        List<byte[]> byteList = Arrays.stream(columnNameStrs).map(String::getBytes).collect(Collectors.toList());
-                        columnNames = byteList.toArray(new byte[0][0]);
-                    } catch (Exception e) {
-                        LOG.error("Cannot parse column headers, error is: {}", e);
-                        throw e;
-                    }
-                }
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private String[] parseLine(Text value, Context context) {
-        try {
-            return csvParser.parseLine(value.toString());
-        } catch (Exception e) {
-            LOG.error("Cannot parse line '{}', error is: {}", value.toString(), e.getMessage());
-            context.getCounter(this.getClass().getSimpleName(), "PARSE_ERRORS").increment(1);
-            return null;
-        }
-    }
-
-    private void writeRow(Context context, String rowKeyStr, String[] fields) throws IOException, InterruptedException {
-        ImmutableBytesWritable rowKey = new ImmutableBytesWritable();
-
-        rowKey.set(rowKeyStr.getBytes());
-        Put put = new Put(rowKey.copyBytes());
-
-        for (int i = 0; i < fields.length; i++) {
-            if (!fields[i].isEmpty()) {
-                try {
-                    put.add(new KeyValue(rowKey.get(), columnFamily, columnNames[i], fields[i].getBytes()));
-                } catch (Exception e) {
-                    LOG.error("Cannot write line '{}'", fields[0], e);
-                    throw e;
-                }
-            }
-
-        }
-
-        // Add last updated column
-        put.add(new KeyValue(rowKey.get(), columnFamily, LAST_UPDATED_BY_COLUMN, LAST_UPDATED_BY_VALUE));
-
-        context.write(rowKey, put);
-    }
-
-}
