+++ project/plugins.sbt
@@ -5,5 +5,6 @@ addSbtPlugin("uk.gov.ons" % "sbt-play-defaults" % "0.1-SNAPSHOT")
 
 addSbtPlugin("uk.gov.ons" % "sbt-code-quality" % "0.1-SNAPSHOT")
 
+addSbtPlugin("uk.gov.ons" % "sbt-code-quality" % "0.1-SNAPSHOT")
