+++ repository-hbase/src/main/java/hbase/load/ExitSignals.java
@@ -1,0 +1,7 @@
+package hbase.load;
+
+public class ExitSignals {
+
+    public static final int SUCCESS = 0;
+    public static final int ERROR = -1;
+}
